Fix: Nettoyage des arguments du champ IntegerField pour le prix

+++ shop/models.py
@@ -11,7 +11,7 @@ class Produit(models.Model):
     categorie = models.ForeignKey(Categorie, on_delete=models.CASCADE, related_name='produits')
     nom = models.CharField(max_length=200)  # Corrigé ici
     description = models.TextField(blank=True)
-    prix = models.DecimalField(max_digits=10, decimal_places=2)
+    prix = models.IntegerField()
     stock = models.IntegerField(default=0)
     disponible = models.BooleanField(default=True)
     image = models.ImageField(upload_to='produits/', blank=True, null=True)
